tests are seperated in two groups. Execution time reduced to 9 minutes with this approach.
